# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Budget OS",
    "version": "0.5.0"
  },
  "components": {
    "securitySchemes": {
      "identityPlatform": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Identity Platform ID token"
      }
    }
  },
  "security": [
    {
      "identityPlatform": []
    }
  ],
  "paths": {
...
    "/api/v1/workspaces/{ws}/dimensions": {
      "get": {
        "operationId": "listDimensions",
        "parameters": [
          {
            "name": "ws",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registry dimensions visible to the workspace"
          }
        }
      },
      "post": {
        "operationId": "createDimension",
        "parameters": [
          {
            "name": "ws",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9_]{1,40}$"
                  },
                  "label": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "dataType": {
                    "type": "string",
                    "enum": [
                      "ENUM",
                      "TEXT",
                      "REFERENCE",
                      "DATE_BUCKET"
                    ]
                  },
                  "icon": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 300
                  },
                  "color": {
                    "type": "string",
                    "maxLength": 32
                  },
                  "allowedParents": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[a-z][a-z0-9_]{1,40}$"
                    },
                    "default": []
                  },
                  "isRequiredForLeaf": {
                    "type": "boolean",
                    "default": false
                  },
                  "sortOrder": {
                    "type": "integer",
                    "default": 0
                  },
                  "workspaceId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  }
                },
                "required": [
                  "key",
                  "label",
                  "dataType",
                  "icon",
                  "workspaceId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created dimension"
          }
        }
      }
    },
    "/api/v1/dimensions/{id}": {
      "patch": {
        "operationId": "updateDimension",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000,
                    "nullable": true
                  },
                  "icon": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 300
                  },
                  "color": {
                    "type": "string",
                    "maxLength": 32,
                    "nullable": true
                  },
                  "allowedParents": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[a-z][a-z0-9_]{1,40}$"
                    }
                  },
                  "isRequiredForLeaf": {
                    "type": "boolean"
                  },
                  "sortOrder": {
                    "type": "integer"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated dimension"
          }
        }
      }
    },
    "/api/v1/dimensions/{id}/values": {
      "post": {
        "operationId": "addDimensionValues",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "values": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 5000,
                    "items": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "pattern": "^[A-Za-z0-9_]{1,80}$"
                        },
                        "label": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200
                        },
                        "parentCode": {
                          "type": "string",
                          "pattern": "^[A-Za-z0-9_]{1,80}$"
                        },
                        "aliases": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "pattern": "^[A-Za-z0-9_]{1,80}$"
                          },
                          "default": []
                        },
                        "externalIds": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "string"
                          },
                          "default": {}
                        }
                      },
                      "required": [
                        "code",
                        "label"
                      ]
                    }
                  }
                },
                "required": [
                  "values"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upserted dimension values"
          }
        }
      }
    },
    "/api/v1/values/{id}": {
      "patch": {
        "operationId": "updateDimensionValue",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "aliases": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[A-Za-z0-9_]{1,80}$"
                    }
                  },
                  "externalIds": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated dimension value"
          }
        }
      }
    },
    "/api/v1/values/{id}/merge": {
      "post": {
        "operationId": "mergeDimensionValues",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fromCode": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9_]{1,80}$"
                  },
                  "intoCode": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9_]{1,80}$"
                  }
                },
                "required": [
                  "fromCode",
                  "intoCode"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Merged dimension value"
          }
        }
      }
    },
    "/api/v1/workspaces/{ws}/hierarchy-templates": {
      "get": {
        "operationId": "listHierarchyTemplates",
        "parameters": [
          {
            "name": "ws",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hierarchy templates"
          }
        }
      },
      "post": {
        "operationId": "saveHierarchyTemplate",
        "parameters": [
          {
            "name": "ws",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "path": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 20,
                    "items": {
                      "type": "string",
                      "pattern": "^[a-z][a-z0-9_]{1,40}$"
                    }
                  },
                  "isDefault": {
                    "type": "boolean",
                    "default": false
                  }
                },
                "required": [
                  "name",
                  "path"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved hierarchy template"
          }
        }
      }
    },
    "/api/v1/assets": {
      "post": {
        "operationId": "uploadIconAsset",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contentType": {
                    "type": "string",
                    "enum": [
                      "image/svg+xml"
                    ]
                  },
                  "svg": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "contentType",
                  "svg"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sanitized SVG icon asset"
          }
        }
      }
    }
  }
}